Multiple issues fixed
The modifications fix multiple issues.  The inclusion of the two new optional variables (primary_dns  and secondary_dns) allow users to optionally supply the ip addresses for the primary and secondary dns servers to statically assign to the main nic of the vm.  This will allow the computer to join the domin in question even if the vm is on a seperate subnet from the domain (the default network adapter that is used in virtual box is nat).  Further, I also fixed the error that appears when attempting to change the name of the computer when joining it to the domain.  The process which renames the computer is actually done in 2 places if the computer_name variable is assigned a value: it is added to the list of arguments that are used to run the "add-computer" command and using the actual rename-computer function as well.  I chose to remove the section which added the "-newname" parameter in the "generate_command_arguments" function.  Now, it works like a charm