Fixes several changes
These changes fix several issues.  Two additional variables can be set in the vagrant file (primary_dns and secondary_dns) which will statically set the primary and secondary dns servers on the vm's nic, which, if configured correctly, can allow the vm to successfully communicate with the domain controller if the vm was unable to without setting the nic's DNS settings.  Also fixed an issue in regards to renaming the machine while domain-joining it.  The only thing I was unable to get perfect is that when destroying the vm, after the vm is removed from the domain, it has to restart prior to actually destroying the vm, but the plugin does work now.