Several issues fixed. (#26)
* Multiple issues fixed

The modifications fix multiple issues.  The inclusion of the two new optional variables (primary_dns  and secondary_dns) allow users to optionally supply the ip addresses for the primary and secondary dns servers to statically assign to the main nic of the vm.  This will allow the computer to join the domin in question even if the vm is on a seperate subnet from the domain (the default network adapter that is used in virtual box is nat).  Further, I also fixed the error that appears when attempting to change the name of the computer when joining it to the domain.  The process which renames the computer is actually done in 2 places if the computer_name variable is assigned a value: it is added to the list of arguments that are used to run the "add-computer" command and using the actual rename-computer function as well.  I chose to remove the section which added the "-newname" parameter in the "generate_command_arguments" function.  Now, it works like a charm

* Fixes several changes

These changes fix several issues.  Two additional variables can be set in the vagrant file (primary_dns and secondary_dns) which will statically set the primary and secondary dns servers on the vm's nic, which, if configured correctly, can allow the vm to successfully communicate with the domain controller if the vm was unable to without setting the nic's DNS settings.  Also fixed an issue in regards to renaming the machine while domain-joining it.  The only thing I was unable to get perfect is that when destroying the vm, after the vm is removed from the domain, it has to restart prior to actually destroying the vm, but the plugin does work now.

* rename computer fix

Fixes issue where the provisioner always tries to rename the vm.

* Optimizing performance

The plugin forced the vm being provisioned to perform several unnecessary restarts, which unnecessarily extends the amount of time the provisioner needs to complete the tasks it needs to.

* last correct

Resetting this back to what it was originally.  No need to modify

* fix travis config

* new version

* Updated Readme